+++ cypress/integration/app.spec.ts
@@ -12,7 +12,7 @@ describe("Basic e2e", function () {
   it("should create new folder, add Study form, upload Study XML file, add Analysis form, add DAC form, and publish folder", () => {
     cy.login()
 
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     // Add folder name & description, navigate to submissions
     cy.newSubmission()