+++ cypress/integration/doiForm.spec.ts
@@ -4,7 +4,7 @@ describe("DOI form", function () {
 
     cy.login()
 
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     // Add folder name & description, navigate to submissions
     cy.newSubmission()