+++ cypress/integration/draft.spec.ts
@@ -3,7 +3,7 @@ describe("Draft operations", function () {
     cy.task("resetDb")
     cy.login()
 
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     // Add folder name & description, navigate to submissions
     cy.newSubmission()