+++ cypress/integration/editForm.spec.ts
@@ -5,7 +5,7 @@ describe("Populate form and render form elements by object data", function () {
     cy.task("resetDb")
     cy.login()
 
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     // Add folder name & description, navigate to submissions
     cy.newSubmission()