+++ cypress/integration/i18n.spec.ts
@@ -31,7 +31,7 @@ describe("Internationalization", function () {
     cy.get("[data-testid='lang-selector']").click()
     cy.get("[data-testid='fi-lang']").contains("Fi").click()
 
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     cy.url().should("include", "/fi/newdraft")
   })