+++ cypress/integration/linkingAccessionIds.spec.ts
@@ -4,7 +4,7 @@ describe("Linking Accession Ids", function () {
     cy.login()
   })
   it("should link correct accessionIds to the related objects", () => {
-    cy.get("button", { timeout: 10000 }).contains("Create Submission").click()
+    cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
     // Add folder name & description, navigate to submissions
     cy.newSubmission()